+++ apis/aws/v1alpha1/appautoscaling_target_types.go
@@ -44,8 +44,8 @@ type AppautoscalingTargetSpec struct {
 
 	ID string `json:"id,omitempty" tf:"id,omitempty"`
 
-	MaxCapacity int    `json:"maxCapacity" tf:"max_capacity"`
-	MinCapacity int    `json:"minCapacity" tf:"min_capacity"`
+	MaxCapacity int64  `json:"maxCapacity" tf:"max_capacity"`
+	MinCapacity int64  `json:"minCapacity" tf:"min_capacity"`
 	ResourceID  string `json:"resourceID" tf:"resource_id"`
 	// +optional
 	RoleArn           string `json:"roleArn,omitempty" tf:"role_arn,omitempty"`
