+++ apis/aws/v1alpha1/backup_vault_types.go
@@ -50,7 +50,7 @@ type BackupVaultSpec struct {
 	KmsKeyArn string `json:"kmsKeyArn,omitempty" tf:"kms_key_arn,omitempty"`
 	Name      string `json:"name" tf:"name"`
 	// +optional
-	RecoveryPoints int `json:"recoveryPoints,omitempty" tf:"recovery_points,omitempty"`
+	RecoveryPoints int64 `json:"recoveryPoints,omitempty" tf:"recovery_points,omitempty"`
 	// +optional
 	Tags map[string]string `json:"tags,omitempty" tf:"tags,omitempty"`
 }
